Extract serialize_val method to a Anyway::Ext::ValueSerializer module
