+++ lib/anyway/env.rb
@@ -1,17 +1,14 @@
 # frozen_string_literal: true
 
+require 'anyway/ext/value_serializer'
 require 'anyway/ext/deep_dup'
 
 module Anyway
   # Parses environment variables and provides
   # method-like access
   class Env
+    include Anyway::Ext::ValueSerializer
     using Anyway::Ext::DeepDup
-
-    # Regexp to detect array values
-    # Array value is a values that contains at least one comma
-    # and doesn't start/end with quote
-    ARRAY_RXP = /\A[^'"].*\s*,\s*.*[^'"]\z/
 
     def initialize
       @data = {}
@@ -48,30 +45,5 @@ module Anyway
     def get_hash(from, name)
       (from[name] ||= {})
     end
-
-    # rubocop:disable Metrics/MethodLength
-    # rubocop:disable Metrics/CyclomaticComplexity
-    def serialize_val(value)
-      case value
-      when ARRAY_RXP
-        value.split(/\s*,\s*/).map(&method(:serialize_val))
-      when /\A(true|t|yes|y)\z/i
-        true
-      when /\A(false|f|no|n)\z/i
-        false
-      when /\A(nil|null)\z/i
-        nil
-      when /\A\d+\z/
-        value.to_i
-      when /\A\d*\.\d+\z/
-        value.to_f
-      when /\A['"].*['"]\z/
-        value.gsub(/(\A['"]|['"]\z)/, '')
-      else
-        value
-      end
-    end
-    # rubocop:enable Metrics/MethodLength
-    # rubocop:enable Metrics/CyclomaticComplexity
   end
 end
